Add YOLO model integrated with central project
