--- OR-YOLOImagesGradlewT/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Thu Oct 13 17:26:13 BOT 2022
-gradle.version=7.4
