--- OR-YOLOImagesGradlewT/build.gradle
@@ -1,25 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'org.example'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    implementation files(System.getProperty("user.dir") + '\\src\\opencv\\build\\java\\opencv-454.jar')
-}
-
-test {
-    useJUnitPlatform()
-}
-
-task runYOLO (dependsOn: 'classes', type: JavaExec) {
-    main = 'src/main/java/org/example/main.java'
-    classpath = sourceSets.main.runtimeClasspath
-}
