+++ README.md
@@ -102,6 +102,4 @@ YOLO is a Object Recogniser Model that can choose between 80 different objects, 
   - Intellij: https://www.jetbrains.com/es-es/idea/download/#section=windows
 
 ## Important Notes
-> the file yolov3.weights must be downloaded from official YOLO web page and must be paste in the folder /OR-YOLOImagesGradlewT/src/main/java/
-
-> the library OpenCV will be into this specific folder, once downloaded it must be paste in the folder /OR-YOLOImagesGradlewT/src/
+> the file yolov3.weights must be downloaded from official YOLO web page and must be paste in the folder /searchobject/src/main/resources/YOLOConfigurationFiles/
