+++ searchobject/build.gradle
@@ -22,6 +22,7 @@ dependencies {
 	implementation 'org.slf4j:slf4j-simple:2.0.3'
 	implementation 'org.slf4j:slf4j-api:2.0.3'
 	implementation 'net.lingala.zip4j:zip4j:1.3.3'
+	implementation files(System.getProperty("user.dir") + '\\src\\main\\resources\\libraries\\opencv\\opencv-454.jar')
 }
 
 tasks.named('test') {
