+++ searchobject/src/main/java/org/jalau/at18/searchobject/modelrecognizer/TypeModelRecognizer.java
@@ -1,7 +1,7 @@
 package org.jalau.at18.searchobject.modelrecognizer;
 
 public enum TypeModelRecognizer {
-    YOLO("yolo"),
+    YOLO("org/jalau/at18/searchobject/model/yolo"),
     SSD("ssd"),
     OBJECTDETECTION("objectdetection");
     private String model;
